agents_interaction_project/requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dataclasses-json==0.6.7
datasets==2.21.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.39
langchain-experimental==0.0.64
langchain-huggingface==0.0.3
langchain-ollama==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.120
llama-cloud==0.0.17
llama-index==0.11.7
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.0
llama-index-core==0.11.7
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.3
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodejs==0.1.1
npm==0.1.1
numpy==1.26.4
ollama==0.3.1
openai==1.44.0
optional-django==0.1.0
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
psutil==6.0.0
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.24.0
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
sqlparse==0.5.1
striprtf==0.0.26
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.11.0
